refactor(types): migrate type definitions for voice messages, downloads, and DOM elements
- Consolidated and migrated type definitions for voice messages, downloads, and DOM elements into dedicated TypeScript files for improved organization and maintainability.
- Updated existing modules to utilize the new type definitions, enhancing type safety across the codebase.
- Removed redundant type definitions from previous locations to streamline the structure and reduce duplication.
