+++ extension/scripts/types/audio.ts
@@ -3,6 +3,23 @@
  * 音訊分析相關類型定義
  */
 
-// TODO: 在 Phase 9.3 中遷移音訊相關類型
-// 暫時導出空物件，避免 import 錯誤
-export {};
+// ================================================
+// 音訊分析相關介面
+// ================================================
+
+/**
+ * 請求元數據介面
+ */
+export interface RequestMetadata {
+  contentType?: string;
+  contentLength?: string;
+  lastModified?: string;
+}
+
+/**
+ * 音訊持續時間訊息介面
+ */
+export interface AudioDurationMessage {
+  url: string;
+  metadata?: RequestMetadata;
+}
