+++ extension/scripts/types/dom.ts
@@ -3,6 +3,14 @@
  * DOM 操作相關類型定義
  */
 
-// TODO: 在 Phase 9.3 中遷移 DOM 相關類型
-// 暫時導出空物件，避免 import 錯誤
-export {};
+// ================================================
+// DOM 操作相關介面
+// ================================================
+
+/**
+ * 語音訊息元素查找結果介面
+ */
+export interface VoiceMessageElementResult {
+  element: Element;
+  type: "slider";
+}
